Update an item with version check

Update an item while ensuring no conflicts with other changes. If the item has been modified since the specified version, the update will fail.

Usage

update_item_version(client, item_key, data, version)

Arguments

client

Zotero API client object

item_key

Item key to update

data

New item data

version

Current version of the item

Value

Updated item data

Examples

if (FALSE) { # \dontrun{
# Update an item
result <- update_item_version(client, 
                             item_key = "ITEM_KEY",
                             data = list(title = "New Title"),
                             version = 123)
} # }